+++ azure-storage-common/ChangeLog.md
@@ -1,4 +1,4 @@
-2017.12 - version 1.0.0
+2018.01 - version 1.0.0
 
 * Removed `ServiceBuilder.php`, moved static builder methods into `BlobRestProxy`, `TableRestProxy`, `QueueRestProxy` and `FileRestProxy`.
 * Moved method `SharedAccessSignatureHelper::generateBlobServiceSharedAccessSignatureToken()` into `BlobSharedAccessSignatureHelper`.