+++ azure-storage-queue/ChangeLog.md
@@ -1,4 +1,4 @@
-2017.12 - version 1.0.0
+2018.01 - version 1.0.0
 
 * Created `QueueSharedAccessSignatureHelper` and moved method `SharedAccessSignatureHelper::generateQueueServiceSharedAccessSignatureToken()` into `QueueSharedAccessSignatureHelper`.
 * Added static builder methods `createQueueService` into `QueueRestProxy`.